Created objectness_score_mask to fit the dimension needed to estimate a valid objectness mask
The transformers prediction heads estimate a objectness score of shape (B, P, 1). To estimate the objectness mask however a shape of (B, P, 2) is needed. Hence, the last prediction head estimates now a objectness_score of shape (B, P, 1) which is needed to estimate the transformes loss functions and an additional objectness_score_mask of shape (B, P, 1).