added transformer eval as scripts/eval_det.py
eval_det.py doesn't use eval_helper.py, but still leads to different results,
need to find root of problem (maybe check the dataset methods -> conversion to
3d boxes...)

changed eval_helper.py a bit to try finding cause of weird ref acc, but to no
avail. current theory, maybe the ref acc is actually like that. scenario: we
predict two (or even more) bboxes for the same object. both are okay, one has
iou 0.78 and the other iou 0.76. then in the reference module it outputs high
logits for both of them (working well, as intended), however the a slightly
higher value for the one with iou 0.76, essentially choosing that one. since
only the highest iou predicted bbox gets assigned as the ref_ground_truth,
this will lead to a "wrong" ref prediction -> ref_acc = 0% (with only this one
pred). however the iou rate of 0.25 and 0.5 will both be 100%, since it
compares the gt_bboox with the predicted_bbox (which still has iou 0.76 even
though that is not the "best" box, i would argue it is still one of the right
ones).

potential ways to change this could be to revise the way the reference gt,
loss and acc are designed. one option would be to set the ref_ground_truth to
one for each predicted bbox that has higher iou than 0.5 + the best box (to
guarantee a gt when all predictions are worse that iou 0.5). then we could
do cross entropy loss and weight the individual bbox losses according to their
iou. (So for a pred_bbox that has iou 0.9 it will be more important for it to
get a high ref prediction, than for one with iou 0.6). the logits would also
need to go through a sigmoid instead of softmax, since more than one value
will be pushed to one (all bboxes that overlap significantly with reference
gt_bbox).